Fix anoubisd upgrade segfault
Subtracting one from a 32-bit unsigned integer with the value zero yields 0xffffffff which is different from -1 in 64 bit mode.
This caused a segfault for a situation that was unforunate but not problematic on 32 bit systems. Fix by only terminating the
path name buffer with NUL if it is actually there.

Aegis-Change: anoubis.10.972.C616 Delta 312
Bug: 1488
TE BLE