Kernel side events for playground operations
The kernel now generates events for playground operations that cannot be performed inside the playground this includes: opening
of special files in the production environment and renaming of production directories. The daemon currently replies to all events
with an	 immediate EPERM. This will be fixed in subsequent changes.

Aegis-Change: anoubis.10.972.C776 Delta 403
Bug: 1510
BLE