Fix and test playground events for directories
The purpose of this change is to fix and test the playground events for directory renames. Additionally, there is a new open
event for unix socket connects of playground processes. While looking into this, I also found and fixed a bug in the
corresponding code in the SFS module (missing path_put). This change also propagate the events to the anoubis CLI/GUI.

Aegis-Change: anoubis.10.972.C781 Delta 408
Bug: 1510
BLE