Fix Use-after Free Bug in anoubisd
This change fixes a use after free Bug in anoubisd. This can lead to wrong tokens that are being recorded in the reply queue of
the policy engine. As a result the real message is never replied to and the process responsible for the event will hang
indefinitely. This could explain some of our timeout issues in tests.

Aegis-Change: anoubis.10.972.C891 Delta 479
BLE