Playground specific policies and many bugfixes
Support app policies that are only applied to playground processes. Rule wizard (if required) is still missing). This change
fixes several other bugs as well:
* Clean up Details visibility code for Application policies
* Fix a segfault if an AppPolicy is deleted and in other situations. The AnTable and AnListCtrl widget now observe their row
provider and remove it if the row provider is deleted.
* replace ACHAT_RC_PENDING return code of acctranmission with ACHAT_RC_OK if the channel has an event that deals with flushing of
the buffer.
* Fix a long standing BUG in anoubis_msg.h
* Add missing event.Skip() calls
* Move Application Details checkboxes into a Panel to allow hiding of individual checkboxes in the details widget.
NOTE: XenPlayground.pm was run and passes. It was not copied into the change due to a conflict with Change 915.

Aegis-Change: anoubis.10.972.C930 Delta 506
BLE