Fix display of playground list
Sometime xanoubis shows up in the playground list instead of the playground binary. Fix by sleeping 1/10 of a second. The use of
vfork is useful but not sufficient. The problem is a race in the daemon that processes the list request from xanoubis before the
exec event from the kernel.

Aegis-Change: anoubis.10.972.C951 Delta 515
Bug: 1541
TE BLE