Implement playground commit task in GUI
This function implements playground commit in the GUI. Several things are still missing but this is a moderately clean and
working state which is why this change should be committed now. Missing things include:
* Error reporting to the user is incomplete
* Column selection does not work yet and default column widths are broken
* Waiting for the playground commit request from the kernel should have a timeout
* Update of the file list after a commit is sometimes delayed until the mouse is moved. This is probably a bug in wxWidgets that
we will ignore.
* The commit task should have the ability to create backups of production files if requested by the user.
* The commit task should have the ability to not commit a file if this would require overwriting of production system files.
This change includes additional bugfixes:
* Grids and tables must detach from their row provider in the destructor.
* event.Skip() is added liberally.
Finally, the error handling in PlaygroundCtrl is simplified and the taskList_ is removed completely..

Aegis-Change: anoubis.10.973.C105 Delta 6
Bug: 1548
BLE