Fix potential message passing hang in anoubisd
The msg_eof function tries to read data into the buffer using fill_buf. This can cause the data to sit in the buffer unnoticed
because after reading the data the file descriptor is no longer ready for reading while there is no consumer of the data that
would be able to process the data.

Aegis-Change: anoubis.10.973.C160 Delta 38
BLE