Playground task tracking fixes
Tasks that are still tracked by the anoubis daemon but that are no longer running should not keep a playground active. This can
happen if a task somehow manages to pass credentials to a file outside the playground. Can happen with ecryptfs. Additionally,
add a missing pe_proc_put.

Aegis-Change: anoubis.10.973.C202 Delta 77
BLE