Force dynamic policy change if UID of a process changes
Whenever we have to consult a policy we first check if the registered UID of the process matches the UID reported in the event.
If both are equal nothing changes, i.e. the cached policy from the process's context is used. However, if the user IDs are
different, we lookup the appropriate policy for the User-ID in the event and path/checksum (the binary, not the context) of the
running process. This effectively enforces a policy context switch for programs that change their UID. However, if  the program
changes its UID back to the original UID, we still use the old context. REGTESTS passed.

Aegis-Change: anoubis.10.973.C118 Delta 102
BLE