Refactor Playground list requests in generic list requests
This change is a preparation for the process browser. The latter will request a list of processes from the anoubis daemon which
is very similar to the playground list request that we current have. Rename a lot of stuff in the protocol layer and its users.
Thus we now have generic list requests and list replies in the protocol that can easily be extended for processes. We simply
introduce a new record type.

Aegis-Change: anoubis.10.973.C219 Delta 103
BLE