Refactor Playground list requests in generic list requests (Part II)
This change is a preparation for the process browser. The latter will request a list of processes from the anoubis daemon which
is very similar to the playground list requests that we current have. Rename a lot of stuff in the anoubis daemon. Thus we now
have generic list requests and list replies in the protocol and in th daemon. These can easily be extended for processes. We
simply have to introduce a new record type.

Aegis-Change: anoubis.10.973.C232 Delta 105
BLE