Unlink Task memory leak fixes
The unlink task can leak memory in at least two cases:
* If a match list is given the constructed path is never freed.
* If a file has several (potential) names for the same device/inode pair we will replace the pointer stored in the hash with a
new pointer but do not free the old value. This is fixed by replacing the roles in the in the unlinkList: The hash key is now the
string pointer and the hash value is the devInodePair. This has the additional benefit that we will try to unlink all names of a
file in one go even if multiple hardlinks exist.

Aegis-Change: anoubis.10.973.C242 Delta 108
TE BLE