fix ALF default rule logging
The handling of default rules in pe_alf.c was incorrect, it used the normal afilt.filtspec struct instead of apndefault. No
compiler warnings because of the union. However this caused logging for default-rules to fail on amd64, on 32bit we seemingly got
lucky with the layout of the union so things mostly still worked. This change fixes the bug.

Aegis-Change: anoubis.10.973.C341 Delta 177