Playground termination notifications
Send notifications whenever a playground is terminated (i.e. the last process in the playground
exits). The event is displayed as an alert next to the anoubis tray icon. If possible, it
includes the process name in the description. Other changes:
* The stat command of the anoubisctl binary should skip non-stat events
* Change u_intXX to uintXX in anoubisd.h
* Set the anoubis protocol version to 6. We are no longer kompatible with older protocol
versions. The user must update anoubisd and xanoubis simultaneously.
* Print playground IDs in hex when dumping messages
* Fix regression tests that have been broken previously or by the protocol version change

Aegis-Change: anoubis.10.973.C401 Delta 220
BLE