Make sure that only a single line can be selected in AnGrids
Using shift, ctrl or mouse dragging, it is possible to select multiple rows in a
Grid. This is bad as AnGrids are used as a flexible replacement for tables. Fix this
by adding an event handler for the range select event that adjust the selection such
that only the row with the grid cursor can be selected once the mouse is released.

Aegis-Change: anoubis.10.973.C423 Delta 233
Bug: 1513
TE BLE