Fix logging for admin rules
An admin allow rule with a log directive can be overwritten by a user allow rule
without a log directive. In this case there is no log message for the event. This is
at least a theoretical security hole. The new rules are:
* The rule ID and priority of the rule that ultimately decided the even is reported.
This is usually a user's allow rule for allowed events.
* The highest log level of all rules involved in the decision is used to log the
event. This includes allow rules from previous priorities as well as SFS continue
rules.

Aegis-Change: anoubis.10.973.C416 Delta 234