Avoid duplicate autostarted GUIs
With KDE saved sessions (and probably others) it is possible to get
several GUIs on login. One from the xanoubis autostart and another from
the saved session. Avoid this by not starting the GUI if another GUI is
already running and the GUI is started with -H.

Aegis-Change: anoubis.10.973.C435 Delta 245
Bug: 1131
TE BLE