saving last playground id to disk and restoring it after system reboot
/ daemon restart

This change is a preparation for the implementation of a playground
cleanup procedure which will remove dead playgrounds after system
reboot / daemon restart. The behaviour implemented in this change leads
to storing the playground id into the file /var/lib/anoubis/last.pgid
whenever a new playground has been created. At daemon restart the id is
being loaded and compared with the one found in the directory structure
in /var/lib/anoubis/playground. The highest of both values will be used
to prevent eventual collisions with possibly incompletely cleaned up
playgrounds (which so far can happen in case of renamed directories
which weren't empty, as the files in them can not be referenced with
the current implementation (that particular problem will be addressed

Aegis-Change: anoubis.10.974.C118 Delta 12
BLE