repaired playground list bug (broken selection states).
This change fixes a bug in the playground overview list by removing the
onRowChange() event handler which caused unneccessary
updateButtonState() calls and broke the internal state of the
playground overview list. The confused list state resulted in invalid /
broken selection states where updateButtonStates() set wrong states to
the buttons due to falsely receving -1 from getFirstSelection() for the
last playground entry in the list. Furthermore the redundant update
calls resulted in unecessary refresh cycles and even performance loss
which grew proportional with the number of playgrounds in the list.
Avoiding those updateButtonState() calls itself does not lead to
invalid states because the function is called anwhere else properly
where necessary ( i.e. in refreshPlaygroundList() ), so that it can be

Aegis-Change: anoubis.10.974.C171 Delta 36
TE BLE