Reset SIGCHLD handler in playground master process.
Fedora-13 (others not checked) does not reset the SIGCHLD handler on
and exec if it is set to SIGIGN. According to the man page this is
allowed but not mandated, i.e. we must deal with either behaviour.
Leaving the child signal as SIGIGN causes problems with Xephyr. Work
around by resetting the SIGCHLD handler in the playground process after
the fork.

Aegis-Change: anoubis.10.974.C205 Delta 53
TE BLE